hw008 (replaced toDouble() -> toFloat(), without optimizations toFloat())

+++ app/src/main/java/com/example/firsthomework/MainActivity.kt
@@ -97,7 +97,7 @@ class MainActivity : AppCompatActivity() {
                     tvResult.text = sum.toString()
                 }
                 "÷" -> {
-                    val sum = enteredNumber / lastNumber.toDouble()
+                    val sum = enteredNumber / lastNumber.toFloat()
                     tvProcess.text = ""
                     tvResult.text = sum.toString()
                 }
